Split Contour UI into it's own implementation file.
